+++ examples/api-auth-jwt-cognito-user-pool/packages/functions/src/private.ts
@@ -1,0 +1,10 @@
+import { APIGatewayProxyHandlerV2WithJWTAuthorizer } from "aws-lambda";
+
+export const main: APIGatewayProxyHandlerV2WithJWTAuthorizer = async (
+  event
+) => {
+  return {
+    statusCode: 200,
+    body: `Hello ${event.requestContext.authorizer.jwt.claims.sub}!`,
+  };
+};