+++ examples/create-sst-dynamo/sst.config.ts
@@ -6,7 +6,7 @@ import { Database } from "./stacks/Database";
 export default {
   config(_input) {
     return {
-      name: "@@app",
+      name: "create-sst-dynamo",
       region: "us-east-1",
     };
   },