+++ packages/sst/src/constructs/NextjsSite.ts
@@ -162,7 +162,7 @@ export interface NextjsSiteProps extends Omit<SsrSiteProps, "nodejs"> {
 }
 
 const LAYER_VERSION = "2";
-const DEFAULT_OPEN_NEXT_VERSION = "2.3.1";
+const DEFAULT_OPEN_NEXT_VERSION = "2.3.5";
 const DEFAULT_CACHE_POLICY_ALLOWED_HEADERS = [
   "accept",
   "rsc",
@@ -194,6 +194,7 @@ export class NextjsSite extends SsrSite {
     sourcemapKey?: string;
   } & ({ regexMatch: string } | { prefixMatch: string }))[];
   private routesManifest?: {
+    basePath: string;
     dynamicRoutes: { page: string; regex: string }[];
     staticRoutes: { page: string; regex: string }[];
     dataRoutes?: { page: string; dataRouteRegex: string }[];
@@ -372,14 +373,14 @@ export class NextjsSite extends SsrSite {
               } as const,
               {
                 cacheType: "server",
-                pattern: "api/*",
+                pattern: this.prefixPattern("api/*"),
                 cfFunction: "serverCfFunction",
                 edgeFunction: "edgeServer",
                 origin: "s3",
               } as const,
               {
                 cacheType: "server",
-                pattern: "_next/data/*",
+                pattern: this.prefixPattern("_next/data/*"),
                 cfFunction: "serverCfFunction",
                 edgeFunction: "edgeServer",
                 origin: "s3",
@@ -393,20 +394,20 @@ export class NextjsSite extends SsrSite {
               } as const,
               {
                 cacheType: "server",
-                pattern: "api/*",
+                pattern: this.prefixPattern("api/*"),
                 cfFunction: "serverCfFunction",
                 origin: "regionalServer",
               } as const,
               {
                 cacheType: "server",
-                pattern: "_next/data/*",
+                pattern: this.prefixPattern("_next/data/*"),
                 cfFunction: "serverCfFunction",
                 origin: "regionalServer",
               } as const,
             ]),
         {
           cacheType: "server",
-          pattern: "_next/image*",
+          pattern: this.prefixPattern("_next/image*"),
           cfFunction: "serverCfFunction",
           origin: "imageOptimizer",
         },
@@ -415,11 +416,13 @@ export class NextjsSite extends SsrSite {
           (item) =>
             ({
               cacheType: "static",
-              pattern: fs
-                .statSync(path.join(sitePath, ".open-next/assets", item))
-                .isDirectory()
-                ? `${item}/*`
-                : item,
+              pattern: this.prefixPattern(
+                fs
+                  .statSync(path.join(sitePath, ".open-next/assets", item))
+                  .isDirectory()
+                  ? `${item}/*`
+                  : item
+              ),
               origin: "s3",
             } as const)
         ),
@@ -429,6 +432,14 @@ export class NextjsSite extends SsrSite {
       },
       buildId: this.getBuildId(),
     });
+  }
+
+  private prefixPattern(pattern: string): string {
+    // Prefix CloudFront distribution behavior path patterns with `basePath` if configured
+    const { basePath } = this.useRoutesManifest();
+    return basePath && basePath.length > 0
+      ? `${basePath.slice(1)}/${pattern}`
+      : pattern;
   }
 
   private createRevalidationQueue() {