+++ packages/sst/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @serverless-stack/cli2
 
+## 2.40.4
+
+### Patch Changes
+
+- [#3659](https://github.com/sst/sst/pull/3659) [`e2a019d80`](https://github.com/sst/sst/commit/e2a019d8035bd47ed6ee34731700dbb1b52c9c5c) Thanks [@ZakkProjects](https://github.com/ZakkProjects)! - Runtime: Fix rust local-lambda build step on windows #
+
+- [`f818a1a31`](https://github.com/sst/sst/commit/f818a1a31c2b9613869fa0cb292540eec6f1c126) Thanks [@fwang](https://github.com/fwang)! - Resource binding customize permissions
+
 ## 2.40.3
 
 ### Patch Changes